# coursework/backend/src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: Krusty Crab CRM API
  description: |
    API для системы управления рестораном "Красти Крабс"
    
    Система предназначена для автоматизации бизнес-процессов ресторана:
    - Управление заказами и клиентами
    - Контроль меню и ингредиентов
    - Управление сотрудниками и рабочими сменами
    - Финансовый учет и отчётность
    - Программы лояльности
    - Мониторинг качества обслуживания
  version: 1.0.0
  contact:
    name: Krusty Crab Support
servers:
  - url: http://localhost:13228/api
    description: Local development server
...
paths:
...
  /auth/me:
    get:
      tags:
        - auth
      summary: Получить информацию о текущем пользователе
      description: Получение информации о текущем аутентифицированном пользователе
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
                  username:
                    type: string
                  userType:
                    type: string
                    enum: [CLIENT, EMPLOYEE]
                  role:
                    type: string
                    nullable: true
                    description: Роль сотрудника (только для EMPLOYEE)
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========== CLIENTS ==========
  # ========== CLIENTS ==========
...
  /orders:
    post:
      tags:
        - orders
      summary: Создать заказ
      description: Создание нового заказа (FR-001, FR-008)
      operationId: placeOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceOrderRequest'
      responses:
        '201':
          description: Заказ создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  orderId:
                    type: integer
                    description: ID созданного заказа
        '400':
          $ref: '#/components/responses/BadRequest'

    get:
      tags:
        - orders
      summary: Получить список заказов
      description: Получение списка заказов (для управляющего)
      operationId: getAllOrders
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Фильтр по статусу
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: clientId
          in: query
          description: Фильтр по клиенту
          schema:
            type: integer
      responses:
        '200':
          description: Список заказов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /orders/{orderId}:
    get:
      tags:
        - orders
      summary: Получить информацию о заказе
      description: Получение детальной информации о заказе (FR-003)
      operationId: getOrderById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Информация о заказе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - orders
      summary: Обновить статус заказа
      description: Обновление статуса заказа (FR-008, FR-010, FR-013)
      operationId: updateOrderStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  $ref: '#/components/schemas/OrderStatus'
      responses:
        '200':
          description: Статус заказа обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /orders/{orderId}/payment-method:
    patch:
      tags:
        - orders
      summary: Обновить способ оплаты заказа
      description: Обновление выбранного способа оплаты для заказа (UC-2)
      operationId: updateOrderPaymentMethod
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - paymentMethod
              properties:
                paymentMethod:
                  $ref: '#/components/schemas/PaymentMethod'
      responses:
        '200':
          description: Заказ обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /orders/{orderId}/items:
    get:
      tags:
        - orders
      summary: Получить позиции заказа
      description: Получение списка позиций (строк) заказа.
      operationId: getOrderItems
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Список позиций заказа
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderItem'
        '404':
          $ref: '#/components/responses/NotFound'
  /orders/items/batch:
    post:
      tags:
        - orders
      summary: Получить позиции нескольких заказов
      description: Батч-загрузка позиций заказов (для UI списков и поиска)
      operationId: getOrderItemsBatch
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderItemsBatchRequest'
      responses:
        '200':
          description: Позиции заказов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderItem'
        '400':
          $ref: '#/components/responses/BadRequest'
  /orders/{orderId}/courier:
    patch:
      tags:
        - orders
      summary: Назначить курьера на заказ
      description: Назначение/снятие курьера для delivery-заказа (UC-2).
      operationId: assignCourierToOrder
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignCourierRequest'
      responses:
        '200':
          description: Заказ обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  # ========== COURIERS ==========
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен доступа. Получите токен через /auth/login/client или /auth/login/employee
  parameters:
...
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: integer
      description: ID заказа
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Courier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        vehicleInfo:
          type: string
        available:
          type: boolean
        busy:
          type: boolean
...
    Order:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        createdByEmployeeId:
          type: integer
          nullable: true
        acceptedByEmployeeId:
          type: integer
          nullable: true
        type:
          $ref: '#/components/schemas/OrderType'
        status:
          $ref: '#/components/schemas/OrderStatus'
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        preparingAt:
          type: string
          format: date-time
          nullable: true
        readyAt:
          type: string
          format: date-time
          nullable: true
        cookingDurationSeconds:
          type: integer
          minimum: 0
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        totalAmount:
          type: number
          format: decimal
        courierId:
          type: integer
          nullable: true
        courier:
          $ref: '#/components/schemas/Courier'
          nullable: true
        deliveredAt:
          type: string
          format: date-time
          nullable: true
        deliveryAddress:
          type: string
          nullable: true
    OrderType:
      type: string
      enum:
        - dine_in
        - takeout
        - delivery
    OrderStatus:
      type: string
      enum:
        - pending
        - confirmed
        - preparing
        - ready
        - delivering
        - delivered
        - completed
        - cancelled
    PlaceOrderRequest:
      type: object
      required:
        - clientId
        - type
        - items
      properties:
        clientId:
          type: integer
        type:
          $ref: '#/components/schemas/OrderType'
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        deliveryAddress:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemRequest'
    OrderItemsBatchRequest:
      type: object
      required:
        - orderIds
      properties:
        orderIds:
          type: array
          minItems: 1
          maxItems: 200
          items:
            type: integer
    OrderItemRequest:
      type: object
      required:
        - menuItemId
        - quantity
      properties:
        menuItemId:
          type: integer
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
          format: decimal
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        menuItemId:
          type: integer
        name:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
          format: decimal
        totalPrice:
          type: number
          format: decimal
        note:
          type: string
    AssignCourierRequest:
      type: object
      required:
        - courierId
      properties:
        courierId:
          type: integer
          nullable: true
...
    PaymentMethod:
      type: string
      enum:
        - cash
        - card
        - online
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string